Add files via upload
This is a quick test of the format to use for the pipeline. I have come into a few errors however, I tried running the same code in an R markdown and had no issues when knitting it. I have googled the errors received and resolved them, however, that only caused more errors. Please let me know if you see anything to fix.